+++ docs/boar-man.rst
@@ -9,8 +9,16 @@ boar.testing
    :undoc-members:
    :show-inheritance:
 
+boar.linting
+------------
+
+.. automodule:: boar.linting
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
 boar.running
--------------
+------------
 
 .. automodule:: boar.running
    :members:
