+++ src/boar/linting.py
@@ -3,16 +3,42 @@ from typing import Any, List, Union
 
 from boar.__init__ import BoarError
 from boar.utils.log import log_lint
-from boar.utils.parse import get_code_execution_counts, remove_output
+from boar.utils.parse import get_code_execution_counts, get_cell_counts, remove_output
 
 
 def lint_notebook(
     notebook_path: Union[str, Path],
-    inline: bool,
-    verbose: Any,
+    inline: bool = False,
+    verbose: Any = True,
     recursion_level: int = 0,
     max_recursion: Union[int, None] = None,
 ) -> List[str]:
+    """Lint notebook.
+
+    Parameters
+    ----------
+    notebook_path : Union[str, Path]
+        Notebook path or notebook directory
+    inline : bool, optional
+        Replace existing notebook with linted version, by default False
+    verbose : Any, optional
+        Verbosity option, by default True
+    recursion_level : int, optional
+        Level of recurssion, by default 0
+        Set to -1000 if you wish to avoid raising Error
+    max_recursion : Union[int, None], optional
+        Depth of directory to explore, by default None
+
+    Returns
+    -------
+    List[str]
+        Posix of notebook that failed
+
+    Raises
+    ------
+    BoarError
+        At list one notebook as failed, the message will list all failed notebooks
+    """
     notebook_path = Path(notebook_path)
     incorrect_files = []
 
@@ -80,7 +106,3 @@ def lint_file(
     file_posix = Path(file_path).as_posix()
     log_lint(file_posix, cell_counts, verbose)
     return file_posix
-
-
-def get_cell_counts(counts):
-    return [(idx+1, count) for idx, count in enumerate(counts) if count is not None]
